+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.0] - 2026-07-08
+### Added
+- **ProxyCommandConfig** to convenient filtering some sets of commands
+
+### Fixed
+- Controls factory for empty tooltips
+
 ## [1.6.0] - 2026-07-08
 ### Changed
 - Updated `common-core` dependency to version **1.3.0**